+++ main/modules/stepper_motor.cpp
@@ -207,14 +207,14 @@ double StepperMotor::position() {
     return static_cast<double>(this->properties.at("position")->integer_value);
 }
 
-void StepperMotor::position(const double position, const double speed, const uint32_t acceleration) {
-    this->position(static_cast<int32_t>(position), static_cast<int32_t>(speed), acceleration);
+void StepperMotor::position(const double position, const double speed, const double acceleration) {
+    this->position(static_cast<int32_t>(position), static_cast<int32_t>(speed), static_cast<uint32_t>(acceleration));
 }
 
 double StepperMotor::speed() {
     return static_cast<double>(this->properties.at("speed")->integer_value);
 }
 
-void StepperMotor::speed(const double speed, const uint32_t acceleration) {
-    this->speed(static_cast<int32_t>(speed), acceleration);
+void StepperMotor::speed(const double speed, const double acceleration) {
+    this->speed(static_cast<int32_t>(speed), static_cast<uint32_t>(acceleration));
 }
